requirement.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
